requirements.txt
Django
coreapi
python_dateutil
djangorestframework
django-webpack-loader
django-cors-headers
django-anymail[sparkpost]
pyyaml
gunicorn
gevent
psycopg2
postgres